Added CarController,cars migration table and CRUD blade files

+++ database/migrations/2023_04_20_204317_create_cars_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('car_image')->nullable();
             $table->integer('horsepower');
             $table->integer('topspeed');
-            $table->integer('acceleration');
+            $table->float('acceleration');
             $table->date('model');
             $table->decimal('price', 9, 2);
             $table->timestamps();
